Adds group member page

+++ app/controllers/api/memberships_controller.rb
@@ -1,9 +1,12 @@
 class Api::MembershipsController < ApplicationController
   def create
     @membership = Membership.new(membership_params)
-    if @membership.save
-      @group = @membership.group
-      render 'api/groups/show'
+    if @membership.user.id != current_user.id
+      render json: ['Invalid credentials'], status: 401
+    elsif @membership.save
+      head :no_content
+      # @group = @membership.group
+      # render 'api/groups/show'
     else
       render json: @membership.errors.full_messages, status: 400
     end
@@ -16,29 +19,18 @@ class Api::MembershipsController < ApplicationController
     elsif @membership.nil?
       render json: ['Membership not found'], status: 404
     elsif @membership.user_id == current_user.id ||
-          @membership.group.organize.id == current_user.id
-      custom_destroy_item
+          @membership.group.organizer.id == current_user.id
+      if @membership.destroy
+        head :no_content
+      else
+        render json: @membership.errors.full_messages, status: 400
+      end
     else
       render json: ['Invalid credentials'], status: 401
     end
-
   end
 
   private
-
-  def custom_destroy_item
-    if @membership.destroy
-      if params[:type] == "Group"
-        @group = @membership.group
-        render 'api/groups/show'
-      elsif params[:type] == "User"
-        @user = @membership.user
-        render 'api/users/show'
-      end
-    else
-      render json: @membership.errors.full_messages, status: 400
-    end
-  end
 
   def membership_params
     params.require(:membership).permit(:group_id, :user_id)
